requirements.txt
absl-py==0.12.0
aiofiles==0.5.0
aiohttp==3.6.2
albumentations==0.5.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
backcall==0.2.0
blis==0.7.4
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
configparser==5.0.2
cycler==0.10.0
cymem==2.0.5
Cython==0.29.22
dataclasses==0.6
decorator==4.4.2
docker-pycreds==0.4.0
effdet==0.2.1
fastai==2.1.4
fastcore==1.3.2
fastprogress==1.0.0
fire==0.4.0
fsspec==0.8.7
future==0.18.2
fvcore==0.1.3.post20210311
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.27.1
google-auth-oauthlib==0.4.3
grpcio==1.36.1
h11==0.9.0
httptools==0.1.1
icedata @ git+git://github.com/airctic/icedata.git@0dcfbdc95cdd4fb765135a0cf4ee991446bed1c2
icevision==0.6.0b1
idna==2.10
imageio==2.9.0
imagesize==1.2.0
imgaug==0.4.0
iopath==0.1.6
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
loguru==0.5.3
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mmpycocotools==12.0.3
multidict==4.7.6
murmurhash==1.0.5
networkx==2.5
nose==1.3.7
numpy==1.19.0
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
packaging==20.9
pandas==1.2.3
parso==0.8.1
pathtools==0.1.2
pathy==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
portalocker==2.2.1
preshed==3.0.5
promise==2.3
prompt-toolkit==3.0.17
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydantic==1.7.3
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
pytorch-lightning==1.2.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
resnest==0.0.6b20210315
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.1
sentry-sdk==1.0.0
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
smart-open==3.0.0
smmap==3.0.5
spacy==3.0.5
spacy-legacy==3.0.1
srsly==2.4.0
starlette==0.13.6
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
thinc==8.0.2
threadpoolctl==2.1.0
tifffile==2021.3.5
timm==0.4.5
torch==1.7.1
torchtext==0.8.1
torchvision==0.8.2
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.11.7
uvloop==0.15.2
wandb==0.10.22
wasabi==0.8.2
wcwidth==0.2.5
websockets==8.1
Werkzeug==1.0.1
yacs==0.1.8
yarl==1.6.3